Renamed 'containers' to 'stores'
