+++ src/api/request.js
@@ -1,4 +1,4 @@
-import store from '../containers/store';
+import store from '../stores/store';
 import superagent from 'superagent';
 
 let token = '';
